% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usdhuf.R
\name{get_usdhuf}
\alias{get_usdhuf}
\title{Get usd/huf rate}
\usage{
get_usdhuf(retried = 0)
}
\arguments{
\item{retried}{number of times failed in the past}
}
\value{
number
}
\description{
Get usd/huf rate
}
